\name{dduCopula-methods}
\docType{methods}
\alias{dduCopula-methods}
\alias{dduCopula,matrix,claytonCopula-method}
\alias{dduCopula,matrix,frankCopula-method}
\alias{dduCopula,matrix,gumbelCopula-method}
\alias{dduCopula,matrix,indepCopula-method}
\alias{dduCopula,matrix,normalCopula-method}
\alias{dduCopula,matrix,tCopula-method}
\alias{dduCopula,matrix,leafCopula-method}
\alias{dduCopula,numeric,claytonCopula-method}
\alias{dduCopula,numeric,frankCopula-method}
\alias{dduCopula,numeric,gumbelCopula-method}
\alias{dduCopula,numeric,indepCopula-method}
\alias{dduCopula,numeric,normalCopula-method}
\alias{dduCopula,numeric,tCopula-method}
\alias{dduCopula,numeric,leafCopula-method}

\title{Methods for Function \code{dduCopula} in Package \pkg{spcopula}}
\description{
 Methods for function \code{dduCopula} in package \pkg{spcopula}
}
\section{Methods}{
\describe{

\item{\code{signature(u = "matrix", copula = "claytonCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "matrix", copula = "frankCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "matrix", copula = "gumbelCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "matrix", copula = "indepCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "matrix", copula = "normalCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "matrix", copula = "tCopula")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(u = "numeric", copula = "claytonCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "frankCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "gumbelCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "indepCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "normalCopula")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "tCopula")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{partial derivatives}
