\name{simulatedTriples}
\alias{triples}
\docType{data}
\title{
annual extreme rainfall triples
}
\description{
The datset contains of the annual extreme rainfall events used in the study:
\dQuote{Joint return periods in hydrology: a critical and practical review focusing on synthetic design hydrograph estimation} by Vandenberghe et al. (2012).
}
\usage{data(simulatedTriples)}
\format{
  The format is:
 peak, duration, volume
}
\references{
\dQuote{Joint return periods in hydrology: a critical and practical review focusing on synthetic design hydrograph estimation} by Vandenberghe et al. (2012).
}
\examples{
data(simulatedTriples)
str(triples)
}
\keyword{datasets}
