\name{spCopDemo}
\alias{lokliktau}
\alias{bestFitTau}
\alias{bins}
\alias{calcKTauLin}
\alias{calcKTauPol} 
\alias{dataSet}
\alias{gevEsti}
\alias{loc}
\alias{loglikTau}
\alias{meanLog}
\alias{scale}
\alias{sdLog}
\alias{shape}
\alias{spCop}
\alias{meuseSpVine}
\docType{data}
\title{
workspace produced in \code{demo(spCopula)}
}
\description{
This workspace is produced by the call \code{demo(spCopula)} and used in a couple of examples throughout the package.
}
\usage{data("spCopDemo")}
\references{
Graeler, B. & E. Pebesma (2011): The pair-copula construction for spatial data: a new approach to model spatial dependency. Poster at: Spatial Statistics 2011 - Mapping global change. Enschede, The Netherlands, 23-25 March 2011. DOI: 10.1016/j.proenv.2011.07.036.}
\examples{
data("spCopDemo")
ls()
}
\keyword{datasets}
