\name{spcopula-package}
\alias{spcopula-package}
\alias{spcopula}
\docType{package}
\title{
copula driven spatial analysis
}
\description{
This package provides a framework to analyse spatial data provided in the format of the \code{\link[sp:sp]{sp}} package with copulas. Additionally, support for calculating multivariate return periods is implemented.
}
\details{
\tabular{ll}{
Package: \tab spcopula\cr
Type: \tab Package\cr
Version: \tab 1.0.67\cr
Date: \tab 2012-10-31\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
Depends: \tab copula, spacetime, VineCopula, methods, lattice, R (>= 2.13.2)\cr
URL: \tab http://r-forge.r-project.org/projects/spcopula/\cr
Collate: \tab Classes.R partialDerivatives.R cqsCopula.R asCopula.R\cr
\tab spcopula.R stcopula.R spatialPreparation.R linkingCDVine.R\cr
\tab BB1copula.R BB6copula.R BB7copula.R BB8copula.R JoeCopula.R\cr
\tab ClaytonGumbelCopula.R vineCopulas.R utilities.R returnPeriods.R\cr
Packaged: \tab 2012-10-31 12:23:06 UTC; b_grae02\cr
Built: \tab R 3.0.0; ; 2013-04-05; windows\cr
}

Index:
\preformatted{
spcopula-package        spatial copula
}

}
\author{
Benedikt Graeler

Maintainer: Benedikt Graeler <ben.graeler@uni-muenster.de>
}
\keyword{ package }
