\name{surClaytonCopula}
\alias{surClaytonCopula}
\alias{r90ClaytonCopula}
\alias{r270ClaytonCopula}
\title{
suvival and rotated Clayton Copulas
}
\description{
these are wrappers to functions from \code{\link{CDVine-package}}
}
\usage{
surClaytonCopula(param)
r90ClaytonCopula(param)
r270ClaytonCopula(param)
}
\arguments{
  \item{param}{
A single parameter defining the Copula.
}
}
\value{
An object of class \code{\linkS4class{surClaytonCopula}}, \code{\linkS4class{r90ClaytonCopula}} or \code{\linkS4class{r270ClaytonCopula}} respectively.
}
\author{
Benedikt Graeler
}
\examples{
persp(surClaytonCopula(1.5),dCopula, zlim=c(0,10))
persp(r90ClaytonCopula(-1.5),dCopula, zlim=c(0,10))
persp(r270ClaytonCopula(-1.5),dCopula, zlim=c(0,10))
}
\keyword{copula}
