\name{invdduCopula-methods}
\docType{methods}
\alias{invdduCopula}
\alias{invdduCopula-methods}
\alias{invdduCopula,numeric,claytonCopula,numeric-method}
\alias{invdduCopula,numeric,frankCopula,numeric-method}
\alias{invdduCopula,numeric,indepCopula,numeric-method}
\alias{invdduCopula,numeric,normalCopula,numeric-method}
\alias{invdduCopula,numeric,spCopula,ANY-method}

\title{Methods for Function \code{invdduCopula} in Package \pkg{spcopula}}
\description{
Methods for function \code{invdduCopula} in package \pkg{spcopula}
}
\section{Methods}{
\describe{

\item{\code{signature(u = "numeric", copula = "claytonCopula", y = "numeric")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "frankCopula", y = "numeric")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "indepCopula", y = "numeric")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "normalCopula", y = "ANY")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{inverse partial derivatives}