\name{invddvCopula-methods}
\docType{methods}
\alias{invddvCopula}
\alias{invddvCopula-methods}
\alias{invddvCopula,numeric,claytonCopula,numeric-method}
\alias{invddvCopula,numeric,frankCopula,numeric-method}
\alias{invddvCopula,numeric,indepCopula,numeric-method}
\alias{invddvCopula,numeric,normalCopula,numeric-method}
\title{Methods for Function \code{invddvCopula} in Package \pkg{spcopula}}
\description{
Methods for function \code{invddvCopula} in package \pkg{spcopula}
}
\section{Methods}{
\describe{

\item{\code{signature(u = "numeric", copula = "claytonCopula", y = "numeric")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "frankCopula", y = "numeric")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "indepCopula", y = "numeric")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(u = "numeric", copula = "normalCopula", y = "ANY")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
\keyword{inverse partial derivatives}