\name{simulatedTriples}
\alias{triples}
\docType{data}
\title{
annual extreme rainfall triples
}
\description{
The dataset contains of the annual extreme rainfall events used in the study:
\dQuote{Graeler, B., M. J. van den Berg, S. Vandenberghe, A. Petroselli, S. Grimaldi, B. De Baets, & N. E. C. Verhoest (2013): Multivariate return periods in hydrology: a critical and practical review focusing on synthetic design hydrograph estimation. Hydrology and Earth System Sciences, 17, 1281-1296.}.
}
\usage{data("simulatedTriples")}
\format{
  The format is:
 peak, duration, volume
}
\references{
Graeler, B., M. J. van den Berg, S. Vandenberghe, A. Petroselli, S. Grimaldi, B. De Baets, & N. E. C. Verhoest (2013): Multivariate return periods in hydrology: a critical and practical review focusing on synthetic design hydrograph estimation. Hydrology and Earth System Sciences, 17, 1281-1296.
}
\examples{
data("simulatedTriples")
str(triples)
}
\keyword{datasets}
