\name{spCopula-package}
\alias{spCopula-package}
\alias{spCopula}
\docType{package}
\title{spatial copula}
\description{This package connects the functionality of copulas with the spatial (and sptaio-temporal) domain. It heavily relies on the packages \code{copula} and \code{sp}.}
\details{
\tabular{ll}{
Package: \tab spCopula\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2011-03-22\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{Benedikt Gräler \email{ben.graeler@uni-muenster.de}}
\references{Some information can be found on my personal homepage \url{http:\\ifgi.uni-muenster.de\graeler} and in the source code. (Will be extended)
}
\keyword{ package }
\usage{Currently there are two new copula classes (asCopula, cqsCopula) that are in many ways compatible with the classes provided in the copula package. A set of partial derivatives accesable through dducopual or ddvcopula with syntax similair to dcopula. Leading to the vine-copula approach, classes and functions on neighbourhoods are present.}
\seealso{ }
\examples{ }
